#50 rename : sns, OAuth 로고 이름 일관성 있게 수정
